shared: Simplify command generator calling
There are three functions within the shared module which generate
various shell commands.  They all take dictionaries as arguments, but
each only uses a single key from those dicts.  This adds unnecessary
overhead for callers - passing in a string is much easier than building
a dictionary for a single value.

We can have the functions expect a string value, but if they get a
dictionary, they will fall back to previous behaviour.  Since this
module is potentially shared outside of this codebase, there may be call
sites which we cannot update to support the new behaviour.

The three call sites within the codebase to these functions are updated
to pass in strings, eliminating the now-unnecessary scaffolding.

NOTE: these are the only call points for these functions in the entire
codebase, and they are all used as input variables for Jinja templates.
It may be more appropriate to do this within the templates themselves,
rather than have such processing being done in two separate places.
This can be done in a future commit.