Simplify ISO builder config handling
We pass boatloads of arguments into the ISO builder object, and keep an
even larger number in the object itself, almost all of which are copied
verbatim out of the existing config blobs.  Instead, let's do the config
handling outside of the builder, put it into an AttributeDict, and
reference items directly from the collected configuration.  The only
downside is that this diff is immense; we change several of the scripts,
and a huge number of lines within the IsoBuild class.

The configs are all collected together in a new method called
`collect_config` in the common module.  It takes the base config (which
up to now has been constant and unmodifiable), overlays the
configuration for the desired release, and further overlays the
command-line arguments, all into the same structure.  The resulting
object is a copy of the individual configs, so can be modified without
affecting the main configurations.  This also opens up the ability to
override some of the global configuration inside each release config.

We had one name collision in combining the groups of config items:  the
base config grabs the architecture of the machine it's running on and
calls it `arch`, and most of the scripts we're modifying also take in an
`--arch` command-line argument, which also lands in `arch`.  It seems to
matter which one we want in the places we use them, so we change the
command-line argument to land in an attribute called `build_arch`.

There were a couple instances we noticed of instance attributes never
being used, which allowed us to delete them.  There were another couple
instances of instance attributes being used only once; we inlined these
definitions and removed the attributes.

This sort of change may be useful for operations which don't use the
IsoBuild, but that would make for an even larger diff.  We can modify
the other classes individually in the future.