IsoBuild: Rearrange conf-file methods to ease readability
These methods are long and difficult to read, because things which
impact any given file are spread over the whole of the method.  We'll
move most things which concern a given file into locality with each
other, which helps readers to follow what is going on.

Writing each file likely belongs in its own method, but that's a bigger
change; at the least, this collects those parts into chunks that we can
more easily spin off into their own methods.