IsoBuild: Move reposcan retrieval where it's used
Within the "do the thing" method for extra ISOs, we grab the `reposcan`
parameter from the image config, pass it into the graft builder, and
never use it again.  We'll move that retrieval into the graft builder
where it's used.  While we're making the change, we'll simplify the
retrieval with a dictionary `.get()`.