IsoBuild: Move repo lists for grafts into graft builder
The repo lists come out of the config object, which is part of the class
and available everywhere.  We already have the key to use for selecting
the repos in a separate call parameter, so there is no need to pass the
repo lists.

We move the retrieval into the graft builder and eliminate the related
call parameter.