IsoBuild: Move graft builder call where the results are used
We call out to the graft builder from the "do the thing" method, only to
pass the result into the template rendering method, and never use it
again.  We'll move the graft builder call near where we actually use the
results.  This allows us to eliminate one of the call parameters of the
graft builder.